WebServer and bug fixes. Fixes #5

+++ AudioStreamingLibCore/AudioStreamingLibCore/audiooutput.h
@@ -48,7 +48,7 @@ private:
     bool m_is_get_very_output_enabled;
     bool m_buffer_requested;
     bool m_play_called;
-    int m_volume;
+    float m_volume;
     float m_sample_align;
     int m_size_to_buffer;
     int m_time_to_buffer;
